Factor "obtain a lock manager" out of algorithms (#94)
More factoring of logic out of #75 - this pulls the logic for obtaining a lock manager (or failing for opaque origins) out of call sites into a new set of steps, taking an environment as input. This sets us up for the later work to support partitioning.

A few other tiny wording/syntax tweaks are included as well, but there are no intended normative behavior changes.